+++ src/pages/profilePage/index.js
@@ -1,11 +1,12 @@
 import WidgetWrapper from "../../components/WidgetWrapper";
 import Navbar from "../../widgets/Navbar";
+import Form from "./form.jsx"
 
 const ProfilePage = () => {
   return (
     <>
       <Navbar />
-      <WidgetWrapper>Profile Page</WidgetWrapper>
+      <Form/>
     </>
   );
 };
